Disallow '.' and '-' in hashtags
So "#a-a" will be parsed as "#a". That matches Instagram's
implementation.

I had originally chosen to allow '.' and '-' in the middle of hashtags
because it made sense if your hashtags were likely to be names of soccer
teams. And since this crate was developed for use at Tonsser, that was
very likely.

However it introduces lots of edge cases such as parsing "#team-name" as
"#team-name", but what about "#f.c."? Should that be "#f.c" or "#f.c."?
If we allow the trailing '.' then what about "Rust is #awesome.". There
you probably don't want to include the '.'